feat: Support for adding `@opentelemetry/instrumentation-runtime-node`
