+++ lib/exporter/ConsoleSpanExporter.js
@@ -4,6 +4,14 @@ const LOG = cds.log('telemetry')
 const path = require('path')
 
 const { ExportResultCode, hrTimeToMilliseconds } = require('@opentelemetry/core')
+const {
+  SEMATTRS_HTTP_TARGET,
+  SEMATTRS_CODE_FILEPATH,
+  SEMATTRS_CODE_LINENO
+  // SEMATTRS_CODE_COLUMN,
+} = require('@opentelemetry/semantic-conventions')
+// REVISIT: SEMATTRS_CODE_COLUMN doesn't yet exist in semantic conventions 1.25
+const SEMATTRS_CODE_COLUMN = 'code.column'
 
 const _padded = v =>
   `${`${v}`.split('.')[0].padStart(3, ' ')}.${(`${v}`.split('.')[1] || '0').padEnd(2, '0').substring(0, 2)}`
@@ -18,25 +26,26 @@ const _span2line = (span, parentStartTime = 0) => {
   let result = `\n  ${_padded(start)} → ${_padded(end)} = ${_padded(duration)} ms`
 
   let name = span.name
-  if (name.match(/^[A-Z]+$/)) name = name + ' ' + span.attributes['http.target']
+  if (name.match(/^[A-Z]+$/)) name = name + ' ' + span.attributes[SEMATTRS_HTTP_TARGET]
   if (name.length > 80) name = name.substring(0, 79) + '…'
 
   result += '  ' + (span.___indent || '') + name
 
   // REVISIT: what is this for?
-  if (span.attributes['code.filepath'] !== undefined) {
+  // REVISIT: SEMATTRS_CODE_COLUMN doesn't yet exist in semantic conventions 1.25
+  if (span.attributes[SEMATTRS_CODE_FILEPATH] !== undefined) {
     if (
       path
-        .normalize(span.attributes['code.filepath'])
+        .normalize(span.attributes[SEMATTRS_CODE_FILEPATH])
         .match(new RegExp(path.normalize(cds.env._home).replaceAll('\\', '\\\\'), 'g')) &&
-      !path.normalize(span.attributes['code.filepath']).match(/node_modules/g)
+      !path.normalize(span.attributes[SEMATTRS_CODE_FILEPATH]).match(/node_modules/g)
     ) {
       result += `: .${path
-        .normalize(span.attributes['code.filepath'])
+        .normalize(span.attributes[SEMATTRS_CODE_FILEPATH])
         .substring(
           path.normalize(cds.env._home).length + 1,
-          path.normalize(span.attributes['code.filepath']).length
-        )}:${span.attributes['code.lineno']}:${span.attributes['code.column']}`
+          path.normalize(span.attributes[SEMATTRS_CODE_FILEPATH]).length
+        )}:${span.attributes[SEMATTRS_CODE_LINENO]}:${span.attributes[SEMATTRS_CODE_COLUMN]}`
     }
   }
