+++ lib/index.js
@@ -1,10 +1,22 @@
 const cds = require('@sap/cds')
 
 const { diag } = require('@opentelemetry/api')
+const { registerInstrumentations } = require('@opentelemetry/instrumentation')
+const { SEMRESATTRS_SERVICE_NAME, SEMRESATTRS_SERVICE_VERSION } = require('@opentelemetry/semantic-conventions')
 
 const tracing = require('./tracing')
 const metrics = require('./metrics')
-const { getDiagLogLevel, getResource } = require('./utils')
+const { getDiagLogLevel, getResource, _require } = require('./utils')
+
+function _getInstrumentations() {
+  const instrumentations = []
+  for (const each of Object.values(cds.env.requires.telemetry.instrumentations)) {
+    const module = _require(each.module)
+    if (!module[each.class]) throw new Error(`Unknown instrumentation "${each.class}" in module "${each.module}"`)
+    instrumentations.push(new module[each.class]({ ...(each.config || {}) }))
+  }
+  return instrumentations
+}
 
 module.exports = function () {
   // set logger and propagate log level
@@ -14,17 +26,22 @@ module.exports = function () {
 
   // REVISIT: better way to make available?
   cds._telemetry = {
-    name: resource.attributes['service.name'],
-    version: resource.attributes['service.version']
+    name: resource.attributes[SEMRESATTRS_SERVICE_NAME],
+    version: resource.attributes[SEMRESATTRS_SERVICE_VERSION]
   }
 
   /*
-   * add tracing
+   * setup tracing
    */
   tracing(resource)
 
   /*
-   * add metrics
+   * setup metrics
    */
   metrics(resource)
+
+  /*
+   * register instrumentations
+   */
+  registerInstrumentations({ instrumentations: _getInstrumentations() })
 }
