+++ lib/tracing/index.js
@@ -3,7 +3,6 @@ const LOG = cds.log('telemetry')
 
 const { trace } = require('@opentelemetry/api')
 const { Resource } = require('@opentelemetry/resources')
-const { registerInstrumentations } = require('@opentelemetry/instrumentation')
 const { BatchSpanProcessor, SimpleSpanProcessor, SamplingDecision } = require('@opentelemetry/sdk-trace-base')
 const { NodeTracerProvider } = require('@opentelemetry/sdk-trace-node')
 
@@ -63,16 +62,6 @@ function _getPropagator() {
   return new core.CompositePropagator({ propagators })
 }
 
-function _getInstrumentations() {
-  const instrumentations = []
-  for (const each of Object.values(cds.env.requires.telemetry.instrumentations)) {
-    const module = _require(each.module)
-    if (!module[each.class]) throw new Error(`Unknown instrumentation "${each.class}" in module "${each.module}"`)
-    instrumentations.push(new module[each.class]({ ...(each.config || {}) }))
-  }
-  return instrumentations
-}
-
 function _getExporter() {
   let {
     kind,
@@ -125,8 +114,6 @@ module.exports = resource => {
     resource = new Resource({}).merge(resource).merge(dtmetadata)
     tracerProvider = new NodeTracerProvider({ resource, sampler: _getSampler() })
     tracerProvider.register({ propagator: _getPropagator() })
-    const instrumentations = _getInstrumentations()
-    registerInstrumentations({ tracerProvider, instrumentations })
   } else {
     LOG._warn && LOG.warn('TracerProvider already initialized by a different module. It will be used as is.')
     tracerProvider = tracerProvider.getDelegate()
