+++ lib/tracing/trace.js
@@ -3,11 +3,12 @@ const cds = require('@sap/cds')
 const otel = require('@opentelemetry/api')
 const { SpanKind, SpanStatusCode, ROOT_CONTEXT } = otel
 const {
+  SEMATTRS_HTTP_TARGET,
   SEMATTRS_CODE_FUNCTION,
   SEMATTRS_CODE_FILEPATH,
   SEMATTRS_CODE_NAMESPACE,
   SEMATTRS_CODE_LINENO,
-  SEMATTRS_CODE_COLUMN,
+  // SEMATTRS_CODE_COLUMN,
   SEMATTRS_DB_SYSTEM,
   SEMATTRS_DB_NAME,
   SEMATTRS_DB_USER,
@@ -18,6 +19,8 @@ const {
   SEMATTRS_DB_OPERATION,
   SEMATTRS_DB_SQL_TABLE
 } = require('@opentelemetry/semantic-conventions')
+// REVISIT: SEMATTRS_CODE_COLUMN doesn't yet exist in semantic conventions 1.25
+const SEMATTRS_CODE_COLUMN = 'code.column'
 
 const { _hrnow } = require('../utils')
 
@@ -49,7 +52,8 @@ function _getParentSpan() {
       // root span gets request attributes
       _setAttributes(parent, _getRequestAttributes())
       if (HRTIME) parent.startTime = cds.context.http?.req?.__hrnow || _hrnow()
-      if (ADJUST_ROOT_NAME && parent.attributes['http.target']) parent.name += ' ' + parent.attributes['http.target']
+      if (ADJUST_ROOT_NAME && parent.attributes[SEMATTRS_HTTP_TARGET])
+        parent.name += ' ' + parent.attributes[SEMATTRS_HTTP_TARGET]
     }
     if (!parent?._is_async) cds.context._otelctx.setValue(cds.context._otelKey, parent)
   }
