+++ package.json
@@ -39,6 +39,7 @@
     "@opentelemetry/exporter-trace-otlp-proto": "^0.52.1",
     "@opentelemetry/host-metrics": "^0.35.0",
     "@opentelemetry/instrumentation": "^0.52.1",
+    "@opentelemetry/instrumentation-runtime-node": "^0.6.0",
     "@sap/cds-mtxs": "^2.0.5",
     "axios": "^1.6.7",
     "chai": "^4.4.1",
